// src/lib/api/types.ts
export type SessionResponse = {
  token: string;
  expiresAt: string;
  userId: string;
  email: string;
  firstName: string;
  lastName: string;
  initials: string;
  baseCurrency: string;
};